+++ deepmd/op/CMakeLists.txt
@@ -17,11 +17,12 @@ if (BUILD_CPP_IF)
         set (EXTRA_LIBS ${EXTRA_LIBS} deepmd_op_cuda)
         target_link_libraries (${LIB_DEEPMD_OP} ${EXTRA_LIBS})
         target_link_libraries (${LIB_DEEPMD_OP} ${CUDA_LIBRARIES})
-    else (USE_CUDA_TOOLKIT)
+    else ()
         add_library(${LIB_DEEPMD_OP} SHARED ${OP_SRC})
-    endif (USE_CUDA_TOOLKIT)
-endif (BUILD_CPP_IF)
+    endif ()
+endif ()
 
+message(STATUS "ORIGIN ${ORIGIN}")
 if (BUILD_PY_IF)
   set(CMAKE_BUILD_WITH_INSTALL_RPATH TRUE)
   set(CMAKE_INSTALL_RPATH $ORIGIN)
@@ -34,10 +35,10 @@ if (BUILD_PY_IF)
     set (EXTRA_LIBS ${EXTRA_LIBS} deepmd_op_cuda)
     target_link_libraries (op_abi ${EXTRA_LIBS})
     target_link_libraries (op_grads ${EXTRA_LIBS})
-  else (USE_CUDA_TOOLKIT)
+  else()
     add_library(op_abi SHARED ${OP_SRC} ${OP_LIB})
     add_library(op_grads SHARED ${OP_GRADS_SRC})
-  endif(USE_CUDA_TOOLKIT)
+  endif()
   message(STATUS ${TensorFlowFramework_LIBRARY})
   target_link_libraries(
     op_abi ${TensorFlowFramework_LIBRARY}
@@ -55,7 +56,7 @@ if (BUILD_PY_IF)
     PROPERTIES
     COMPILE_FLAGS ${OP_CXX_FLAG}
     )
-endif (BUILD_PY_IF)
+endif()
 
 if (BUILD_CPP_IF)
   install(TARGETS ${LIB_DEEPMD_OP}		DESTINATION lib/)
