+++ deepmd/op/_soft_min_force_grad.py
@@ -4,16 +4,17 @@ Gradients for soft min force
 """
 
 from tensorflow.python.framework import ops
-from deepmd.env import op_grads_module
 
-     
+from deepmd.train.env import op_grads_module
+
+
 @ops.RegisterGradient("SoftMinForce")
-def _soft_min_force_grad_cc (op, grad):    
-    net_grad = op_grads_module.soft_min_force_grad (grad, 
-                                                    op.inputs[0], 
-                                                    op.inputs[1], 
-                                                    op.inputs[2], 
-                                                    op.inputs[3], 
-                                                    n_a_sel = op.get_attr("n_a_sel"),
-                                                    n_r_sel = op.get_attr("n_r_sel"))
+def _soft_min_force_grad_cc(op, grad):
+    net_grad = op_grads_module.soft_min_force_grad(grad,
+                                                   op.inputs[0],
+                                                   op.inputs[1],
+                                                   op.inputs[2],
+                                                   op.inputs[3],
+                                                   n_a_sel=op.get_attr("n_a_sel"),
+                                                   n_r_sel=op.get_attr("n_r_sel"))
     return [net_grad, None, None, None]
